+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [0.31.0] - 2020-09-28
 
+### Added
+- Modal: Added Alert with overlay.
+
 ## [0.30.0] - 2020-09-14
 
 ### Added